Makefile update. Modules can now be compiled into chameleon. Modified some chameleon functions to follow the std c lib
git-svn-id: http://forge.voodooprojects.org/svn/chameleon/trunk@891 b5af796a-00a8-11df-919b-ffff7a100b7d

+++ i386/libsa/string.c
@@ -108,7 +108,7 @@ void bzero(void * dst, size_t len)
 #define tolower(c)     ((int)((c) & ~0x20))
 #define toupper(c)     ((int)((c) | 0x20))
 
-int strlen(const char * s)
+size_t strlen(const char * s)
 {
 	int n = 0;
 	while (*s++) n++;
